Add new build job and refactor unit test job
Introduce a new `new-build` job in `dotnet.yml` for building and publishing versions, running on `windows-latest`. This job depends on the `new-version` job and only runs if it is successful and the version is valid.

Rename the `test` job to `invoke-unit-tests` and update it to run on `ubuntu-latest`. This job now depends on both `new-build` and `new-version` jobs and only runs if `new-build` is successful.

Refactor `invoke-unit-tests` to include new steps: checkout repository, list files, and invoke unit tests using `g4-api/g4-github-actions/actions/test-dotnet-project@v1` with specific conditions and parameters. Remove previous steps for setting up .NET, restoring dependencies, building, and running tests.

Add environment variables `buildVersion`, `projectName`, and `projectType` to the `invoke-unit-tests` job.