Update dotnet.yml: enhance build and add unit test job
Modified the `run` command for the `msbuild` task to include the `${{ env.solutionName }}.sln` file, build configuration, and output path for the binaries directory. Added a new job named `invoke-unit-tests` that runs on `ubuntu-latest`, executes only if the `new-build` job was successful, depends on the completion of both `new-build` and `new-version` jobs, and includes default settings for the `run` command.